refactor: memberService 공통 메서드 추출 #32
